+++ packages/components/src/components/Tab.vue
@@ -25,6 +25,7 @@ export default {
       type: Boolean,
       default: false,
     },
+    reference: String,
   },
 
   data() {